refactor: Consolidate heading autolink config in dedicated file
